+++ Cesium3DTilesSelection/include/Cesium3DTilesSelection/TilesetContentManager.h
@@ -1,9 +1,8 @@
 #pragma once
 
-#include "RasterOverlayUpsampler.h"
-
 #include <Cesium3DTilesSelection/CesiumIonTilesetContentLoaderFactory.h>
 #include <Cesium3DTilesSelection/RasterOverlayCollection.h>
+#include <Cesium3DTilesSelection/RasterOverlayUpsampler.h>
 #include <Cesium3DTilesSelection/Tile.h>
 #include <Cesium3DTilesSelection/TileContent.h>
 #include <Cesium3DTilesSelection/TilesetContentLoader.h>
@@ -17,6 +16,7 @@
 #include <CesiumUtility/ReferenceCounted.h>
 
 #include <vector>
+
 
 namespace Cesium3DTilesSelection {