+++ Cesium3DTilesSelection/src/TilesetContentLoader.cpp
@@ -1,5 +1,4 @@
-#include "TilesetContentManager.h"
-
+#include <Cesium3DTilesSelection/TilesetContentManager.h>
 #include <Cesium3DTilesSelection/TileContent.h>
 #include <Cesium3DTilesSelection/TileLoadResult.h>
 #include <Cesium3DTilesSelection/TilesetContentLoader.h>