+++ CesiumVectorOverlays/src/GeoJsonDocumentRasterOverlay.cpp
@@ -12,8 +12,8 @@
 #include <CesiumGeospatial/Projection.h>
 #include <CesiumGltf/ImageAsset.h>
 #include <CesiumRasterOverlays/CreateRasterOverlayTileProviderParameters.h>
-#include <CesiumRasterOverlays/GeoJsonDocumentRasterOverlay.h>
-#include <CesiumRasterOverlays/Library.h>
+#include <CesiumVectorOverlays/GeoJsonDocumentRasterOverlay.h>
+#include <CesiumVectorOverlays/Library.h>
 #include <CesiumRasterOverlays/RasterOverlay.h>
 #include <CesiumRasterOverlays/RasterOverlayLoadFailureDetails.h>
 #include <CesiumRasterOverlays/RasterOverlayTile.h>
@@ -47,11 +47,12 @@ using namespace CesiumGeometry;
 using namespace CesiumGeospatial;
 using namespace CesiumUtility;
 using namespace CesiumVectorData;
+using namespace CesiumRasterOverlays;
 
 // We won't generate any quadtree nodes past this depth.
 const uint32_t DEPTH_LIMIT = 8;
 
-namespace CesiumRasterOverlays {
+namespace CesiumVectorOverlays {
 
 namespace {
 /**